pyproject.toml
[project]
name = "capstone2-be"
version = "0.1.0"
description = ""
authors = [
    {name = "KHM",email = "maxkang0328@naver.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "geoalchemy2 (>=0.18.4,<0.19.0)",
    "redis (>=7.4.0,<8.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)"
]

[tool.poetry]
# 패키징(배포용 wheel/sdist) 대신, 의존성 관리 용도로만 Poetry를 사용합니다.
# 그래서 `poetry install` 시 "No file/folder found for package ..." 에러가 나지 않게 합니다.
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
